backend/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
torch==2.0.0
torchvision==0.15.0
matplotlib==3.7.1
contourpy==1.0.7
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.74
scipy==1.11.0
